refactor: Extract API client, parsers, and services from monolith (Phase 1)
Reduced fantasy_football_multi_league.py from 2,675 → 2,154 lines (19% reduction)

Extracted modules:
- src/api/yahoo_client.py (140 lines)
  - yahoo_api_call() with rate limiting and caching
  - refresh_yahoo_token() with auto-retry
  - Centralized token management

- src/parsers/yahoo_parsers.py (193 lines)
  - parse_team_roster() - extract roster from Yahoo API response
  - parse_yahoo_free_agent_players() - extract free agent data

- src/services/reddit_service.py (265 lines)
  - analyze_reddit_sentiment() - enhanced analyzer with fallback
  - _analyze_reddit_sentiment_fallback() - basic PRAW implementation

Benefits:
✅ Easier to find code (clear module organization)
✅ Easier to test (isolated, pure functions)
✅ Easier to modify (change API logic without touching handlers)
✅ Reduced intimidation factor (largest file now 2,154 lines vs 2,675)

All 18 MCP tools verified working - no breaking changes.

Phase 2 (handlers extraction) deferred for future session.
