+++ fantasy_football_multi_league.py
@@ -6,23 +6,17 @@ Fantasy Football MCP Server - Multi-League Support
 import asyncio
 import json
 import os
-import socket
 from typing import Any, Awaitable, Callable, Dict, List, Optional
 
-import aiohttp
 from dotenv import load_dotenv
 from mcp.server import Server
 from mcp.server.stdio import stdio_server
 from mcp.types import TextContent, Tool
 
-# Reddit sentiment analysis imports
-try:
-    import praw
-    from textblob import TextBlob
-
-    REDDIT_AVAILABLE = True
-except ImportError:
-    REDDIT_AVAILABLE = False
+# Import extracted modules
+from src.api import get_access_token, refresh_yahoo_token, set_access_token, yahoo_api_call
+from src.parsers import parse_team_roster, parse_yahoo_free_agent_players
+from src.services import analyze_reddit_sentiment
 
 # Import rate limiting and caching utilities
 from yahoo_api_utils import rate_limiter, response_cache
@@ -33,119 +27,15 @@ DRAFT_AVAILABLE = True
 # Load environment
 load_dotenv()
 
-# Configuration
-YAHOO_ACCESS_TOKEN = os.getenv("YAHOO_ACCESS_TOKEN")
-YAHOO_API_BASE = "https://fantasysports.yahooapis.com/fantasy/v2"
-
-# Reddit configuration
-REDDIT_CLIENT_ID = os.getenv("REDDIT_CLIENT_ID")
-REDDIT_CLIENT_SECRET = os.getenv("REDDIT_CLIENT_SECRET")
-REDDIT_USERNAME = os.getenv("REDDIT_USERNAME")
+# Initialize access token in the API module
+if os.getenv("YAHOO_ACCESS_TOKEN"):
+    set_access_token(os.getenv("YAHOO_ACCESS_TOKEN"))
 
 # Create server instance
 server = Server("fantasy-football")
 
 # Cache for leagues
 LEAGUES_CACHE = {}
-
-
-async def yahoo_api_call(
-    endpoint: str, retry_on_auth_fail: bool = True, use_cache: bool = True
-) -> dict:
-    """Make Yahoo API request with rate limiting, caching, and automatic token refresh."""
-    global YAHOO_ACCESS_TOKEN
-
-    # Check cache first (if enabled)
-    if use_cache:
-        cached_response = await response_cache.get(endpoint)
-        if cached_response is not None:
-            return cached_response
-
-    # Apply rate limiting
-    await rate_limiter.acquire()
-
-    url = f"{YAHOO_API_BASE}/{endpoint}?format=json"
-    headers = {"Authorization": f"Bearer {YAHOO_ACCESS_TOKEN}", "Accept": "application/json"}
-
-    connector = aiohttp.TCPConnector(family=socket.AF_INET)
-    async with aiohttp.ClientSession(connector=connector, trust_env=True) as session:
-        async with session.get(url, headers=headers) as response:
-            if response.status == 200:
-                data = await response.json()
-                # Cache successful response
-                if use_cache:
-                    await response_cache.set(endpoint, data)
-                return data
-            elif response.status == 401 and retry_on_auth_fail:
-                # Token expired, try to refresh
-                refresh_result = await refresh_yahoo_token()
-                if refresh_result.get("status") == "success":
-                    # Token refreshed, retry the API call with new token
-                    return await yahoo_api_call(
-                        endpoint, retry_on_auth_fail=False, use_cache=use_cache
-                    )
-                else:
-                    # Refresh failed, raise the original error
-                    text = await response.text()
-                    raise Exception(f"Yahoo API auth failed and token refresh failed: {text[:200]}")
-            else:
-                text = await response.text()
-                raise Exception(f"Yahoo API error {response.status}: {text[:200]}")
-
-
-async def refresh_yahoo_token() -> dict:
-    """Refresh the Yahoo access token using the refresh token."""
-    global YAHOO_ACCESS_TOKEN
-
-    client_id = os.getenv("YAHOO_CONSUMER_KEY")
-    client_secret = os.getenv("YAHOO_CONSUMER_SECRET")
-    refresh_token = os.getenv("YAHOO_REFRESH_TOKEN")
-
-    if not all([client_id, client_secret, refresh_token]):
-        return {"status": "error", "message": "Missing credentials in environment"}
-
-    token_url = "https://api.login.yahoo.com/oauth2/get_token"
-
-    data = {
-        "client_id": client_id,
-        "client_secret": client_secret,
-        "refresh_token": refresh_token,
-        "grant_type": "refresh_token",
-    }
-
-    try:
-        connector = aiohttp.TCPConnector(family=socket.AF_INET)
-        async with aiohttp.ClientSession(connector=connector, trust_env=True) as session:
-            async with session.post(token_url, data=data) as response:
-                if response.status == 200:
-                    token_data = await response.json()
-                    new_access_token = token_data.get("access_token")
-                    new_refresh_token = token_data.get("refresh_token", refresh_token)
-                    expires_in = token_data.get("expires_in", 3600)
-
-                    # Update global token
-                    YAHOO_ACCESS_TOKEN = new_access_token
-
-                    # Update environment
-                    os.environ["YAHOO_ACCESS_TOKEN"] = new_access_token
-                    if new_refresh_token != refresh_token:
-                        os.environ["YAHOO_REFRESH_TOKEN"] = new_refresh_token
-
-                    return {
-                        "status": "success",
-                        "message": "Token refreshed successfully",
-                        "expires_in": expires_in,
-                        "expires_in_hours": round(expires_in / 3600, 1),
-                    }
-                else:
-                    error_text = await response.text()
-                    return {
-                        "status": "error",
-                        "message": f"Failed to refresh token: {response.status}",
-                        "details": error_text[:200],
-                    }
-    except Exception as e:
-        return {"status": "error", "message": f"Error refreshing token: {str(e)}"}
 
 
 async def discover_leagues() -> dict[str, dict[str, Any]]:
@@ -315,182 +205,6 @@ async def get_user_team_key(league_key: Optional[str]) -> Optional[str]:
     """Get the user's team key in a specific league (legacy function for compatibility)."""
     team_info = await get_user_team_info(league_key)
     return team_info["team_key"] if team_info else None
-
-
-def parse_team_roster(data: dict) -> list[dict]:
-    """Extract a simple roster list from Yahoo team data."""
-    roster: list[dict] = []
-    team = data.get("fantasy_content", {}).get("team", [])
-
-    for item in team:
-        if isinstance(item, dict) and "roster" in item:
-            roster_data = item["roster"]
-            players = None
-            if isinstance(roster_data, dict):
-                players = roster_data.get("0", {}).get("players")
-                if not players:
-                    players = roster_data.get("players")  # Direct players key
-                if not players and "roster" in roster_data:
-                    players = roster_data["roster"].get("players")
-            if not players:
-                print(
-                    f"DEBUG: No players in roster_data for item {item.get('roster', {}).keys() if isinstance(item.get('roster'), dict) else type(item.get('roster'))}"
-                )
-                continue
-
-            for key, pdata in players.items():
-                if key == "count" or not isinstance(pdata, dict) or "player" not in pdata:
-                    continue
-
-                player_array = pdata["player"]
-                if not isinstance(player_array, list):
-                    continue
-
-                info: dict[str, Any] = {}
-
-                # Helper to robustly extract selected position regardless of structure
-                def _extract_position(selected_position_obj: Any) -> Optional[str]:
-                    if not selected_position_obj:
-                        return None
-                    if isinstance(selected_position_obj, dict):
-                        # Direct form: {"position": "WR"}
-                        if "position" in selected_position_obj:
-                            return selected_position_obj.get("position")
-                        # Keyed form: {"0": {"position": "WR"}, "count": 1}
-                        for k, v in selected_position_obj.items():
-                            if k == "count":
-                                continue
-                            if isinstance(v, dict) and "position" in v:
-                                return v["position"]
-                    elif isinstance(selected_position_obj, list):
-                        for entry in selected_position_obj:
-                            if isinstance(entry, dict) and "position" in entry:
-                                return entry["position"]
-                    return None
-
-                def _scan_container(container: Any) -> None:
-                    if not isinstance(container, dict):
-                        return
-                    name_dict = container.get("name")
-                    if isinstance(name_dict, dict) and "full" in name_dict:
-                        info["name"] = name_dict.get("full")
-                    # Status, default will be set later if absent
-                    if "status" in container:
-                        info["status"] = container.get("status", "OK")
-                    # Prefer selected_position when available
-                    if "selected_position" in container:
-                        pos = _extract_position(container.get("selected_position"))
-                        if pos:
-                            info["position"] = pos
-                    # Fallback to display position if selected_position not found
-                    if "position" not in info and "display_position" in container:
-                        info["position"] = container.get("display_position")
-
-                    if "team" not in info:
-                        team_value: Optional[str] = None
-
-                        # Direct keys commonly present in Yahoo responses
-                        for key in (
-                            "editorial_team_abbr",
-                            "team_abbr",
-                            "team_abbreviation",
-                            "editorial_team_full_name",
-                            "editorial_team_name",
-                        ):
-                            value = container.get(key)
-                            if isinstance(value, str) and value.strip():
-                                team_value = value
-                                break
-
-                        # Nested team objects occasionally hold the abbreviation/name
-                        if not team_value and isinstance(container.get("team"), dict):
-                            team_container = container["team"]
-                            for nested_key in ("abbr", "abbreviation", "name", "nickname"):
-                                nested_value = team_container.get(nested_key)
-                                if isinstance(nested_value, str) and nested_value.strip():
-                                    team_value = nested_value
-                                    break
-
-                        if team_value:
-                            info["team"] = team_value
-
-                for element in player_array:
-                    if isinstance(element, dict):
-                        _scan_container(element)
-                    elif isinstance(element, list):
-                        for sub in element:
-                            _scan_container(sub)
-
-                if info:
-                    if "status" not in info:
-                        info["status"] = "OK"
-                    roster.append(info)
-
-    return roster
-
-
-def parse_yahoo_free_agent_players(data: dict) -> list[dict]:
-    """Extract free agent/waiver players from Yahoo data, similar to team roster."""
-    players: list[dict] = []
-    league = data.get("fantasy_content", {}).get("league", [])
-
-    # Find players section (typically league[1]["players"])
-    if len(league) > 1 and isinstance(league[1], dict) and "players" in league[1]:
-        players_data = league[1]["players"]
-
-        for key, pdata in players_data.items():
-            if key == "count" or not isinstance(pdata, dict) or "player" not in pdata:
-                continue
-
-            player_array = pdata["player"]
-            if not isinstance(player_array, list):
-                continue
-
-            info: dict[str, Any] = {}
-
-            def _scan_free_agent(container: Any) -> None:
-                if not isinstance(container, dict):
-                    return
-                # Name
-                name_dict = container.get("name")
-                if isinstance(name_dict, dict) and "full" in name_dict:
-                    info["name"] = name_dict.get("full")
-                # Position
-                if "display_position" in container:
-                    info["position"] = container.get("display_position")
-                # Team
-                if "editorial_team_abbr" in container:
-                    info["team"] = container["editorial_team_abbr"]
-                elif "team" in container and isinstance(container["team"], dict):
-                    info["team"] = container["team"].get("abbr") or container["team"].get(
-                        "abbreviation"
-                    )
-                # Ownership
-                if "ownership" in container and isinstance(container["ownership"], dict):
-                    info["owned_pct"] = container["ownership"].get("ownership_percentage", 0)
-                    info["weekly_change"] = container["ownership"].get("weekly_change", 0)
-                if "percent_owned" in container:
-                    info["owned_pct"] = container["percent_owned"]
-                # Injury
-                if "status" in container:
-                    info["injury_status"] = container["status"]
-                if "status_full" in container:
-                    info["injury_detail"] = container["status_full"]
-                # Bye
-                if "bye_weeks" in container:
-                    info["bye"] = container["bye_weeks"].get("week", "N/A")
-
-            for element in player_array:
-                if isinstance(element, dict):
-                    _scan_free_agent(element)
-                elif isinstance(element, list):
-                    for sub in element:
-                        _scan_free_agent(sub)
-
-            if info and info.get("name"):
-                players.append(info)
-
-    return players
 
 
 async def get_waiver_wire_players(
@@ -730,241 +444,6 @@ async def get_all_teams_info(league_key: str) -> list[dict]:
     except Exception:
         return []
 
-
-async def analyze_reddit_sentiment(
-    players: list[str], time_window_hours: int = 48
-) -> dict[str, Any]:
-    """
-    Analyze Reddit sentiment for fantasy football players using enhanced analyzer.
-    Used for Start/Sit decisions based on community consensus.
-    """
-    try:
-        # Import enhanced Reddit analyzer
-        from src.agents.reddit_analyzer import RedditSentimentAgent
-
-        # Create mock settings for the analyzer
-        class MockSettings:
-            mcp_server_version = "1.0.0"
-
-        # Initialize enhanced analyzer
-        analyzer = RedditSentimentAgent(MockSettings())
-
-        # Analyze sentiment with enhanced error handling
-        if len(players) == 1:
-            # Single player analysis
-            result = await analyzer.analyze_player_sentiment(players[0], time_window_hours)
-
-            # Convert to legacy format for compatibility
-            return {
-                "players": players,
-                "analysis_type": "single",
-                "time_window_hours": time_window_hours,
-                "player_data": {
-                    players[0]: {
-                        "sentiment_score": result.overall_sentiment or 0.0,
-                        "consensus": result.consensus or "MIXED",
-                        "posts_analyzed": result.posts_analyzed,
-                        "comments_analyzed": result.comments_analyzed,
-                        "injury_mentions": result.injury_mentions,
-                        "hype_score": result.hype_score,
-                        "top_comments": result.top_comments[:3],
-                        "status": result.status,
-                        "confidence": result.confidence,
-                        "fallback_used": result.fallback_used,
-                    }
-                },
-                "enhanced_analyzer": True,  # Flag to indicate enhanced version
-            }
-        else:
-            # Multiple player comparison
-            # TODO: Implement compare_players if needed; fallback to single analysis
-            comparison = {
-                "results": {
-                    p: await analyzer.analyze_player_sentiment(p, time_window_hours)
-                    for p in players
-                }
-            }
-
-            # Convert results to legacy format
-            player_data = {}
-            for player_name, result in comparison.get("results", {}).items():
-                player_data[player_name] = {
-                    "sentiment_score": result.overall_sentiment or 0.0,
-                    "consensus": result.consensus or "MIXED",
-                    "posts_analyzed": result.posts_analyzed,
-                    "comments_analyzed": result.comments_analyzed,
-                    "injury_mentions": result.injury_mentions,
-                    "hype_score": result.hype_score,
-                    "top_comments": result.top_comments[:3],
-                    "status": result.status,
-                    "confidence": result.confidence,
-                    "fallback_used": result.fallback_used,
-                }
-
-            return {
-                "players": players,
-                "analysis_type": "comparison",
-                "time_window_hours": time_window_hours,
-                "player_data": player_data,
-                "recommendation": comparison.get("recommendation", {}),
-                "confidence": comparison.get("confidence", 0),
-                "successful_analyses": comparison.get("successful_analyses", 0),
-                "total_players": comparison.get("total_players", len(players)),
-                "timestamp": comparison.get("timestamp", ""),
-                "enhanced_analyzer": True,  # Flag to indicate enhanced version
-            }
-
-        # Clean up
-        await analyzer.cleanup()
-
-    except ImportError as e:
-        # Fallback to basic implementation if enhanced analyzer not available
-        return await _analyze_reddit_sentiment_fallback(
-            players, time_window_hours, f"Enhanced analyzer unavailable: {e}"
-        )
-    except Exception as e:
-        # Fallback to basic implementation on any error
-        return await _analyze_reddit_sentiment_fallback(
-            players, time_window_hours, f"Enhanced analyzer failed: {e}"
-        )
-
-
-async def _analyze_reddit_sentiment_fallback(
-    players: list[str], time_window_hours: int = 48, error_reason: str = ""
-) -> dict[str, Any]:
-    """
-    Fallback Reddit sentiment analysis using basic implementation.
-    Used when enhanced analyzer is unavailable.
-    """
-    if not REDDIT_AVAILABLE:
-        return {
-            "error": "Reddit analysis not available. Install 'praw' and 'textblob' packages.",
-            "fallback_reason": error_reason,
-        }
-
-    if not REDDIT_CLIENT_ID or not REDDIT_CLIENT_SECRET:
-        return {"error": "Reddit API credentials not configured", "fallback_reason": error_reason}
-
-    try:
-        # Initialize Reddit client
-        reddit = praw.Reddit(
-            client_id=REDDIT_CLIENT_ID,
-            client_secret=REDDIT_CLIENT_SECRET,
-            user_agent=f'fantasy-football-mcp:v1.0 by /u/{REDDIT_USERNAME or "unknown"}',
-        )
-
-        results = {
-            "players": players,
-            "analysis_type": "comparison" if len(players) > 1 else "single",
-            "time_window_hours": time_window_hours,
-            "player_data": {},
-            "fallback_used": True,
-            "fallback_reason": error_reason,
-        }
-
-        subreddits = ["fantasyfootball", "DynastyFF", "Fantasy_Football", "nfl"]
-
-        for player in players:
-            player_sentiments = []
-            total_posts = 0
-            total_engagement = 0
-            injury_mentions = 0
-            relevant_comments = []
-
-            # Search across subreddits
-            for subreddit_name in subreddits:
-                try:
-                    subreddit = reddit.subreddit(subreddit_name)
-                    posts = list(subreddit.search(player, time_filter="week", limit=5))
-
-                    for post in posts:
-                        total_posts += 1
-                        total_engagement += post.score + post.num_comments
-
-                        # Analyze sentiment
-                        text = f"{post.title} {post.selftext[:500] if post.selftext else ''}"
-                        blob = TextBlob(text)
-                        sentiment_obj = blob.sentiment
-                        sentiment = getattr(sentiment_obj, "polarity", 0.0)
-                        player_sentiments.append(sentiment)
-
-                        # Check for injuries
-                        injury_keywords = [
-                            "injured",
-                            "injury",
-                            "out",
-                            "doubtful",
-                            "questionable",
-                            "IR",
-                        ]
-                        if any(keyword.lower() in text.lower() for keyword in injury_keywords):
-                            injury_mentions += 1
-
-                        # Get top comments
-                        if post.score > 10:
-                            relevant_comments.append(
-                                {
-                                    "text": post.title[:100],
-                                    "score": post.score,
-                                    "sentiment": sentiment,
-                                }
-                            )
-                except Exception:
-                    continue
-
-            # Calculate metrics
-            avg_sentiment = (
-                sum(player_sentiments) / len(player_sentiments) if player_sentiments else 0
-            )
-
-            # Determine consensus
-            if avg_sentiment > 0.1:
-                consensus = "START"
-            elif avg_sentiment < -0.1:
-                consensus = "SIT"
-            else:
-                consensus = "MIXED"
-
-            # Calculate hype score (combination of sentiment and engagement)
-            hype_score = ((avg_sentiment + 1) / 2) * min(total_engagement / 100, 1.0)
-
-            results["player_data"][player] = {
-                "sentiment_score": round(avg_sentiment, 3),
-                "consensus": consensus,
-                "posts_analyzed": total_posts,
-                "total_engagement": total_engagement,
-                "injury_mentions": injury_mentions,
-                "hype_score": round(hype_score, 3),
-                "top_comments": sorted(relevant_comments, key=lambda x: x["score"], reverse=True)[
-                    :3
-                ],
-            }
-
-        # Add comparison recommendation if multiple players
-        if len(players) > 1:
-            sorted_players = sorted(
-                results["player_data"].items(),
-                key=lambda x: x[1]["sentiment_score"] + x[1]["hype_score"],
-                reverse=True,
-            )
-
-            results["recommendation"] = {
-                "start": sorted_players[0][0],
-                "sit": [p[0] for p in sorted_players[1:]],
-                "confidence": min(
-                    abs(
-                        sorted_players[0][1]["sentiment_score"]
-                        - sorted_players[-1][1]["sentiment_score"]
-                    )
-                    * 100,
-                    100,
-                ),
-            }
-
-        return results
-
-    except Exception as e:
-        return {"error": f"Reddit analysis failed: {str(e)}"}
 
 
 @server.list_tools()
