+++ src/services/__init__.py
@@ -1,0 +1,5 @@
+"""Services for external integrations."""
+
+from .reddit_service import analyze_reddit_sentiment
+
+__all__ = ["analyze_reddit_sentiment"]
